+++ packages/create-app/templates/js-jsx/src/app.jsx
@@ -1,0 +1,9 @@
+export const app = {
+  render() {
+    return (
+      <h1>
+        <Message1 />, <Message2 />, <Message3 />!
+      </h1>
+    )
+  },
+}