+++ packages/create-app/templates/js-jsx/src/store/index.js
@@ -1,0 +1,11 @@
+import { createStore } from "@inglorious/web"
+
+import { message } from "../message/message"
+import { entities } from "./entities"
+import { middlewares } from "./middlewares"
+
+export const store = createStore({
+  types: { message },
+  entities,
+  middlewares,
+})