+++ packages/create-app/templates/js-jsx/vite.config.js
@@ -11,10 +11,7 @@ const __dirname = path.dirname(__filename)
 
 // https://vite.dev/config/
 export default defineConfig({
-  // @see https://github.com/vitejs/vite/issues/1973
-  define: { "process.env": {} },
-
-  plugins: [minifyTemplateLiterals(), , ViteImageOptimizer(), jsx()],
+  plugins: [minifyTemplateLiterals(), ViteImageOptimizer(), jsx()],
 
   resolve: {
     alias: {