+++ packages/create-app/templates/ts-jsx/src/store/entities.ts
@@ -1,0 +1,16 @@
+import type { AppState } from "../../types"
+
+export const entities: AppState = {
+  message1: {
+    type: "message",
+    who: "world",
+  },
+  message2: {
+    type: "message",
+    who: "wide",
+  },
+  message3: {
+    type: "message",
+    who: "web",
+  },
+}