+++ packages/create-app/templates/ts-jsx/src/store/middlewares.ts
@@ -1,0 +1,9 @@
+import { createDevtools, type Middleware } from "@inglorious/web"
+
+import type { AppEntity, AppState } from "../../types"
+
+export const middlewares: Middleware<AppEntity, AppState>[] = []
+
+if (import.meta.env.DEV) {
+  middlewares.push(createDevtools<AppEntity, AppState>().middleware)
+}