+++ packages/create-app/templates/ts-jsx/vite.config.ts
@@ -14,7 +14,7 @@ export default defineConfig({
   // @see https://github.com/vitejs/vite/issues/1973
   define: { "process.env": {} },
 
-  plugins: [minifyTemplateLiterals(), , ViteImageOptimizer(), jsx()],
+  plugins: [minifyTemplateLiterals(), ViteImageOptimizer(), jsx()],
 
   resolve: {
     alias: {