+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/create-game
 
+## 0.4.0
+
+### Minor Changes
+
+- Allow JSX syntax when scaffolding app
+
 ## 0.3.5
 
 ### Patch Changes